moved openai api request to chat component
